pyproject.toml
[project]
name = "fin-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.4.0",
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-structlog>=10.0.0",
    "django-stubs>=5.2.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-stubs>=3.16.0",
    "drf-spectacular>=0.28.0",
    "pandas>=2.3.3",
    "pydantic>=2.11.7",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.1.1",
    "redis>=5.2.0",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "celery-types>=0.23.0",
    "django-types>=0.21.0",
    "djangorestframework-types>=0.9.0",
    "pyrefly>=0.42.3",
    "ruff>=0.12.1",
    "types-python-dateutil>=2.9.0.20250516",
    "types-requests>=2.32.4.20250611",
]


[tool.mypy]
explicit_package_bases = true
packages = "src"
namespace_packages = true
mypy_path = ["src"]
allow_redefinition = true
ignore_missing_imports = true
exclude = ["gunicorn.conf.py", "manage.py", "/migrations/*"]
plugins = ["pydantic.mypy"]

# from https://blog.wolt.com/engineering/2021/09/30/professional-grade-mypy-configuration/
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = false

disallow_incomplete_defs = true
disallow_untyped_decorators = false # default
disallow_any_unimported = false     # handled by Ruff

[tool.django-stubs]
django_settings_module = "fin_manager.settings"
strict_settings = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "fin_manager.settings"
python_files = ["test_*.py"]
pythonpath = ["src"]

[tool.pyrefly]
project-includes = ["src"]
search-path = ["src"]
ignore-missing-imports = ["*"]
